Move MessageData to consts
